carrello iniziale

+++ src/Notebook.java
@@ -1,5 +1,5 @@
 public class Notebook extends ClasseDispositivi{
-    public Notebook(String device, String brand, String model, String description, int display, int storage, double purchase) {
+    public Notebook(String device, String brand, String model, String description, double display, double storage, double purchase) {
         super(device, brand, model, description, display, storage, purchase);
     }
 }
