+++ src/Tablet.java
@@ -1,5 +1,5 @@
 public class Tablet extends ClasseDispositivi{
-    public Tablet(String device, String brand, String model, String description, int display, int storage, double purchase) {
+    public Tablet(String device, String brand, String model, String description, double display, double storage, double purchase) {
         super(device, brand, model, description, display, storage, purchase);
     }
 }
